requirements.txt
antlr4-python3-runtime==4.8
anyio==3.3.4
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
colorlog==6.6.0
configparser==5.1.0
cycler==0.11.0
docker-pycreds==0.4.0
fastapi==0.70.0
fonttools==4.28.1
gitdb==4.0.9
GitPython==3.1.24
h11==0.12.0
hydra-colorlog==1.1.0
hydra-core==1.1.1
idna==3.3
importlib-resources==5.4.0
kiwisolver==1.3.2
matplotlib==3.5.0
numpy==1.21.4
omegaconf==2.1.1
packaging==21.3
pathtools==0.1.2
Pillow==8.4.0
promise==2.3
protobuf==3.19.1
psutil==5.8.0
pydantic==1.8.2
pyparsing==3.0.6
python-dateutil==2.8.2
PyYAML==6.0
requests==2.26.0
sentry-sdk==1.5.0
setuptools-scm==6.3.2
shortuuid==1.0.8
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
starlette==0.16.0
subprocess32==3.5.4
termcolor==1.1.0
tomli==1.2.2
typing_extensions==4.0.0
urllib3==1.26.7
uvicorn==0.15.0
wandb==0.12.7
yaspin==2.1.0
zipp==3.6.0
